% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{afm_processor}
\alias{afm_processor}
\title{Processor for additive factorization machine}
\usage{
afm_processor(term, data, output_dim, param_nr, controls)
}
\description{
Processor for additive factorization machine
}
