% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{ahofm_processor}
\alias{ahofm_processor}
\title{Processor for additive higher-order factorization machine}
\usage{
ahofm_processor(term, data, output_dim, param_nr, controls)
}
\description{
Processor for additive higher-order factorization machine
}
