% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fac_processor}
\alias{fac_processor}
\title{Processor for factor terms with many levels}
\usage{
fac_processor(term, data, output_dim, param_nr, controls = NULL)
}
\description{
Processor for factor terms with many levels
}
