% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fz_processor}
\alias{fz_processor}
\title{Processor for factorization of two or three-level factor interactions}
\usage{
fz_processor(term, data, output_dim, param_nr, controls = NULL)
}
\description{
Processor for factorization of two or three-level factor interactions
}
