% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_partial_effect_latent}
\alias{get_partial_effect_latent}
\title{Partial Predictions Method for safareg}
\usage{
get_partial_effect_latent(
  object,
  names = NULL,
  return_matrix = FALSE,
  which_param = 1,
  newdata = NULL,
  latentdim,
  ...
)
}
\arguments{
\item{object, names, return_matrix, which_param, newdata}{see \code{?get_partial_effect}}

\item{latentdim}{number of latent dimension for AFM or AHOFM}
}
\description{
Partial Predictions Method for safareg
}
