% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{hofm_processor}
\alias{hofm_processor}
\title{Processor for higher-order factorization machine}
\usage{
hofm_processor(term, data, output_dim, param_nr, controls)
}
\description{
Processor for higher-order factorization machine
}
