% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{interaction_processor}
\alias{interaction_processor}
\title{Processor for interactions of factor terms with many levels}
\usage{
interaction_processor(term, data, output_dim, param_nr, controls = NULL)
}
\description{
Processor for interactions of factor terms with many levels
}
