% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.safareg}
\alias{coef.safareg}
\title{Extract layer weights / coefficients from model}
\usage{
\method{coef}{safareg}(object, which_param = 1, type = NULL, ...)
}
\arguments{
\item{object}{a safareg model}

\item{which_param}{integer, indicating for which distribution parameter
coefficients should be returned (default is first parameter)}

\item{type}{either NULL (all types of coefficients are returned),
"linear" for linear coefficients or "smooth" for coefficients of 
smooth terms}

\item{...}{not used}
}
\description{
Extract layer weights / coefficients from model
}
