% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{safareg}
\alias{safareg}
\title{Initialize a Factorized Structured Regression Model}
\usage{
safareg(
  ...,
  additional_processors = list(fac = fac_processor, facz = fz_processor, vc =
    am_processor, vfacz = vf_processor)
)
}
\arguments{
\item{...}{see \code{deepregression}}

\item{additional_processor}{defines how processors of this package are named}
}
\value{
a model of class \code{safareg} and \code{deepregression}
}
\description{
Initialize a Factorized Structured Regression Model
}
\details{
The names of the processors can be choosen freely, but
note that methods for the class \code{safareg} might depend on this
naming convention.
}
