% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{vc_processor}
\alias{vc_processor}
\title{Processor for varying coefficient terms with many factor levels}
\usage{
vc_processor(term, data, output_dim, param_nr, controls)
}
\description{
Processor for varying coefficient terms with many factor levels
}
