% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{vf_processor}
\alias{vf_processor}
\title{Processor for varying coefficient terms with factorization of two factor levels}
\usage{
vf_processor(term, data, output_dim, param_nr, controls = NULL)
}
\description{
Processor for varying coefficient terms with factorization of two factor levels
}
