% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{am_processor}
\alias{am_processor}
\title{Processor for varying coefficient terms with one factor level using array mode}
\usage{
am_processor(term, data, output_dim, param_nr, controls)
}
\description{
Processor for varying coefficient terms with one factor level using array mode
}
